Removed DownloadFile functionality and updated remote config to marshall to a yaml file
